requirements.txt
python-docx>=1.1.0
PyYAML==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
chainlit>=1.0.0
docling-core>=2.0.0
pdf2image>=1.16.3
